Added html structure for sellers profile pages + moved reusable components to a different file (cards with product and artist information)
