+++ group06-handcrafted-haven/src/app/ui/artist/products-by-artist.tsx
@@ -1,0 +1,14 @@
+import { ProductCard } from "@/app/ui/cards";
+
+export default async function ProductsByArtist() {
+  return (
+    <div>
+      <h2>Artist's Products</h2>
+      {/* Display all of the artists products */}
+      <ProductCard />
+      <ProductCard />
+      <ProductCard />
+      <ProductCard />
+    </div>
+  );
+}
